app/requirements.txt
flask
pymysql
mysql-connector-python

